feat: Refactor registration components and update import paths
