+++ resources/js/Pages/Account Creation/Register.jsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import { Head, useForm } from '@inertiajs/react';
 import { Header, Footer } from '@/Components';
-import Step1BasicDetails from './Step1BasicDetails';
-import Step2EducationDetails from './Step2EducationDetails';
-import Step3GameDetails from './Step3GameDetails';
-import Step4AccountCredentials from './Step4AccountCredentials';
+import Step1BasicDetails from './components/Step1BasicDetails.jsx';
+import Step2EducationDetails from './components/Step2EducationDetails.jsx';
+import Step3GameDetails from './components/Step3GameDetails.jsx';
+import Step4AccountCredentials from './components/Step4AccountCredentials.jsx';
 import webBg2025 from './assets/webbg2025.png';
 import './register.css';
 
@@ -198,7 +198,7 @@ const Register = () => {
                 <div className="register-main-bg">
 
 
-                    <img src={webBg2025} className="background-image-register" alt="Web Background" />
+                    {/*<img src={webBg2025} className="background-image-register" alt="Web Background" />*/}
