+++ resources/js/Pages/Account Creation/components/Step1BasicDetails.jsx
