+++ resources/js/Pages/Account Creation/components/Step3GameDetails.jsx
@@ -1,6 +1,6 @@
 // Step3GameDetails.jsx
 import React, { useState, useRef } from "react";
-import MLLogin from '../MLLoginApi/MLLogin';
+import MLLogin from '../../MLLoginApi/MLLogin.jsx';
 
 
 const Step3GameDetails = ({
@@ -15,22 +15,22 @@ const Step3GameDetails = ({
   const [userInfo, setUserInfo] = useState(null);
 
   const loginRef = useRef();
-  
+
   const handleLoginInfo = (info) => {
     console.log('Received Login Info:', info);
-    setUserInfo(info); 
+    setUserInfo(info);
     info ? setIsBlurred(false) : setIsBlurred(true);
     if (info?.code === 0) {
       const data = info.data;
-     
+
       formData.ign = data.name || '';
       formData.level = data.level || '';
       formData.country = data.reg_country || '';
       formData.userId = data.roleId || '';
       formData.serverId = data.zoneId || '';
       formData.avatar = data.avatar || '';
-    
-      
+
+
       const checkMlIdExists = async (ml_id) => {
         const response = await fetch(`/check-ml-id?ml_id=${encodeURIComponent(ml_id)}`);
         const data = await response.json();
@@ -47,10 +47,10 @@ const Step3GameDetails = ({
         formData.serverId = '';
         formData.avatar = '';
         setTimeout(() => {
-          
+
           setIsBlurred(false);
         }, 1000);
-       
+
       }else{
         setErrorMessage("");
         console.log("1");
@@ -58,7 +58,7 @@ const Step3GameDetails = ({
         setIsBlurred(true);
         setIsLoading(false);
       }
-      
+
     }else{
       setIsLoading(false);
     }
@@ -99,7 +99,7 @@ const Step3GameDetails = ({
     <div className="">
       <h1 className="title-register">CREATE MSL ACCOUNT</h1>
       <h2 className="subtitle-register">MLBB DETAILS</h2>
-      
+
       {/* Dynamic Progress Bar for Step 3 */}
       {(() => {
         const requiredFields = [
@@ -143,8 +143,8 @@ const Step3GameDetails = ({
                 </svg>
               </div>
               <div className="electric-wrapper">
-                <div onClick={() => { 
-                  handleLoginClick() 
+                <div onClick={() => {
+                  handleLoginClick()
                   }}
                   className="neon-button text-sm md:text-base px-4 md:px-8 py-2 md:py-3 flex items-center justify-center gap-2">
                   {isLoading ? (
@@ -159,9 +159,9 @@ const Step3GameDetails = ({
                     <span>Connect</span>
                   )}
                 </div>
-                <MLLogin 
-                  ref={loginRef} 
-                  onLoginInfo={handleLoginInfo} 
+                <MLLogin
+                  ref={loginRef}
+                  onLoginInfo={handleLoginInfo}
                   onLoginClose={() => {
                     setIsLoading(false);
                   }}
@@ -239,7 +239,7 @@ const Step3GameDetails = ({
                 value={formData.ign}
                 onChange={handleInputChange}
                 className="input-field-register"
-                
+
               disabled/>
             </div>
             <div className="input-group-register right-side-register">
