+++ resources/js/Pages/Account Creation/register.css
@@ -17,7 +17,6 @@
   padding: 40px 20px;
     position: relative;
     z-index: 1;
-
 }
 
 .background-image-register {
